Uploaded esp01_DHT11grafanaV5
This update introduces a new folder for the latest version of the project. On the RaspberryPi side, I’ve upgraded to InfluxDB 2.X. This change streamlines my data gathering process by eliminating the need for an MQTT bridge. As a result, I can now gather data more efficiently and effectively.